+++ installer/templates/doctor.exs
@@ -4,8 +4,8 @@
   # {Medic.Checks.Chromedriver, :chrome_installed??},
   # {Medic.Checks.Chromedriver, :unquarantined?},
   # {Medic.Checks.Chromedriver, :versions_match?},
-  {Medic.Checks.Direnv, :envrc_file_exists?},
-  {Medic.Checks.Direnv, :has_all_keys?},
+  # {Medic.Checks.Direnv, :envrc_file_exists?},
+  # {Medic.Checks.Direnv, :has_all_keys?},
   {Medic.Checks.Asdf, :plugin_installed?, ["postgres"]},
   {Medic.Checks.Asdf, :package_installed?, ["postgres"]},
   {Medic.Checks.Hex, :local_hex_installed?},
@@ -18,6 +18,6 @@
   # {Medic.Checks.Postgres, :running?},
   # {Medic.Checks.Postgres, :correct_version_running?},
   # {Medic.Checks.Postgres, :role_exists?},
-  # {Local.Checks.Postgres, :correct_data_directory?},
+  # {Medic.Checks.Postgres, :correct_data_directory?},
   # {Medic.Checks.Postgres, :database_exists?, ["my_app_dev"]}
 ]