+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 1.8.1
+
 - `mix.gen.medic` template for `doctor.exs` has fewer checks enabled.
 - More Postgres checks accept a `:remedy` option.